flyio_kube_db/requirements.txt
dnspython==2.2.0
psycopg2==2.9.3
